townley-platform/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
gunicorn==22.0.0
sqlalchemy==2.0.31
pyodbc==5.2.0
pydantic==2.8.0
pydantic-settings==2.5.2
email-validator==2.2.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
prometheus-fastapi-instrumentator==7.0.0
alembic==1.13.2
python-dotenv==1.0.1
structlog==24.1.0
redis==5.0.8
rq==1.16.2
python-multipart==0.0.9
